requirements.txt
# iOS Automation Framework - Python Dependencies
# ============================================

# HTTP Requests
requests>=2.31.0

# Appium Client
Appium-Python-Client>=3.0.0

# Testing Framework
pytest>=7.4.0
pytest-html>=3.2.0
pytest-xdist>=3.5.0
pytest-rerunfailures>=12.0
pytest-base-url>=2.0

# Allure Report
allure-pytest>=2.13.0

# Data Processing
PyYAML>=6.0
jsonpath-ng>=1.6.0

# Mock & Test Utilities
responses>=0.23.0
pytest-mock>=3.11.0
faker>=19.0.0

# Logging & Utils
loguru>=0.7.0
colorama>=0.4.6

# Performance Testing
locust>=2.20.0

# Code Quality
black>=23.0.0
flake8>=6.1.0

# Configuration
python-dotenv>=1.0.0
configparser>=5.3.0

# Async Support
aiohttp>=3.8.0
asyncio-pool>=0.6.0

# Screenshot & Image
Pillow>=10.0.0
